requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
Pillow==10.0.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
whitenoise==6.5.0
